working with curry function
