+++ utils.js
@@ -12,9 +12,6 @@ const askQuestion = function(query) {
     }))
 }
 
-const getPlayers = async () => {
-    const players = await askQuestion("Ingrese nombres de jugadores separados por comas: ")
-    return players.split(",")
-}
 
-module.exports = {askQuestion, getPlayers}
+
+module.exports = {askQuestion}
